I've updated your dungeon character.
Some issues to discuss with Gary. As it's written, only the character with the greatest speed attacks (always). second, in the test bellow, the damage of get_attack_range is a constant value when it should be a range of values (i.e. between 40 to 80; instead of 70 constantly). also, please ask how to apply this to the hero and monster child classes and how to implement special skills.